Internals

Transducers.dcollect — Function

dcollect.([xf,] grouped(key, collection))

Collect each group as a Vector using Distributed.jl.

The first optional argument xf is a transducer.

Example

julia> using LazyGroupBy
       using Transducers

julia> dcollect.(grouped(isodd, [0, 7, 3]))
Transducers.GroupByViewDict{Bool,Array{Int64,1},…} with 2 entries:
  false => [0]
  true  => [7, 3]

Transducers.dreduce — Function

dreduce.(op, [xf,] grouped(key, collection); [init])
dreduce.(os::OnlineStat, [xf,] grouped(key, collection); [init])

The first argument is either a reducing step function or an OnlineStat. The second optional argument xf is a transducer.

Examples

julia> using LazyGroupBy
       using Transducers

julia> dreduce.(+, grouped(isodd, [0, 7, 3, 1, 5, 9, 4, 3, 0, 5]))
Transducers.GroupByViewDict{Bool,Int64,…} with 2 entries:
  false => 4
  true  => 33

Transducers.tcollect — Function

tcollect.([xf,] grouped(key, collection))

Collect each group as a Vector using multiple threads. See also collect.(grouped(key, collection)).

The first optional argument xf is a transducer.

Example

julia> using LazyGroupBy
       using Transducers

julia> tcollect.(grouped(isodd, [0, 7, 3]))
Transducers.GroupByViewDict{Bool,Array{Int64,1},…} with 2 entries:
  false => [0]
  true  => [7, 3]